+++ src/components/recipe_list/RecipeList.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { FlatList } from 'react-native';
 import { NavigationParams, NavigationScreenProp, NavigationState } from 'react-navigation';
 
-import RecipeCard from '../../components/ui-components/recipeCard';
+import RecipeCard from '../ui-components/RecipeCard';
 
 interface RecipeListProps {
   recipes: any;
