+++ src/components/ui-components/RecipeCard.tsx
@@ -4,7 +4,7 @@ import { Button } from 'react-native-elements';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
 import IconFA from 'react-native-vector-icons/FontAwesome';
 
-import { Black, Gray, GrayLighter, White } from '../../styles/config/colors';
+import { Black, Gray, GrayLighter, White } from '../../styles/config/Colors';
 import { NavigationParams, NavigationScreenProp, NavigationState } from 'react-navigation';
 
 interface RecipeCardProps {
