+++ src/screens/home/RecipeListScreen.tsx
@@ -3,9 +3,9 @@ import { View, StyleSheet, SafeAreaView } from 'react-native';
 import { SearchBar } from 'react-native-elements';
 import { NavigationParams, NavigationScreenProp, NavigationState } from 'react-navigation';
 
-import { Gray, Transparent, White } from '../../styles/config/colors';
+import { Gray, Transparent, White } from '../../styles/config/Colors';
 
-import RecipeList from '../../components/recipe_list/recipeList';
+import RecipeList from '../../components/recipe_list/RecipeList';
 import NavHeader from '../../components/ui-components/NavHeader';
 
 interface RecipeScreenState {
