+++ src/screens/onboarding/OnboardingScreen.tsx
@@ -4,7 +4,7 @@ import { Button } from 'react-native-elements';
 import ViewPagerType from '@react-native-community/viewpager';
 
 const ViewPager = require('@react-native-community/viewpager'); // eslint-disable-line @typescript-eslint/no-var-requires
-import { Black } from '../../styles/config/colors';
+import { Black } from '../../styles/config/Colors';
 import { name } from '../../../app.json';
 import AsyncStorage from '@react-native-community/async-storage';
