+++ src/styles/Styles.tsx
@@ -1,0 +1,9 @@
+// COMPONENTS
+
+// CONFIG
+import * as Colors from './config/Colors';
+import * as Spacing from './config/Spacing';
+import * as Typography from './config/Fonts';
+import * as Mixins from './config/Mixins';
+
+export { Typography, Spacing, Colors, Mixins };
