--- src/styles/index.tsx
@@ -1,9 +1,0 @@
-// COMPONENTS
-
-// CONFIG
-import * as Colors from './config/colors';
-import * as Spacing from './config/spacing';
-import * as Typography from './config/font';
-import * as Mixins from './config/mixins';
-
-export { Typography, Spacing, Colors, Mixins };
